pyproject.toml
[project]
name = "resume-to-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "docx2pdf>=0.1.8",
    "python-dotenv>=1.2.1",
]
